git-svn-id: https://svn.jfire.org/svn/main/trunk/org.nightlabs.jfire.trade.ui@19097 18c61ebf-f106-0410-9e1e-81fb82415b8f

+++ src/org/nightlabs/jfire/trade/ui/transfer/wizard/CreditCardPage.java
@@ -163,7 +163,7 @@ public class CreditCardPage extends WizardHopPage
 			@Override
 			public void widgetSelected(org.eclipse.swt.events.SelectionEvent e) {
 				paymentData.setExpiryMonth(
-						((Integer)expiryMonthList.get(expiryMonth.getSelectionIndex())).intValue());
+						(expiryMonthList.get(expiryMonth.getSelectionIndex())).intValue());
 			}
 		});
 
@@ -171,7 +171,7 @@ public class CreditCardPage extends WizardHopPage
 			@Override
 			public void widgetSelected(org.eclipse.swt.events.SelectionEvent e) {
 				paymentData.setExpiryYear(
-						((Integer)expiryYearList.get(expiryYear.getSelectionIndex())).intValue());
+						(expiryYearList.get(expiryYear.getSelectionIndex())).intValue());
 			}
 		});
